+++ pkgs/test_core/lib/src/runner/configuration.dart
@@ -311,6 +311,8 @@ class Configuration {
     required Map<PlatformSelector, SuiteConfiguration>? onPlatform,
     required bool? ignoreTimeouts,
     required Timeout? suiteLoadTimeout,
+    Hook? preRun,
+    Hook? postRun,
 
     // Test-level configuration
     required Timeout? timeout,
@@ -367,6 +369,8 @@ class Configuration {
         onPlatform: onPlatform,
         ignoreTimeouts: ignoreTimeouts,
         suiteLoadTimeout: suiteLoadTimeout,
+        preRun: preRun,
+        postRun: postRun,
 
         // Test-level configuration
         timeout: timeout,
@@ -431,6 +435,8 @@ class Configuration {
     Map<PlatformSelector, SuiteConfiguration>? onPlatform,
     bool? ignoreTimeouts,
     Timeout? suiteLoadTimeout,
+    Hook? preRun,
+    Hook? postRun,
 
     // Test-level configuration
     Timeout? timeout,
@@ -484,6 +490,8 @@ class Configuration {
     onPlatform: onPlatform,
     ignoreTimeouts: ignoreTimeouts,
     suiteLoadTimeout: suiteLoadTimeout,
+    preRun: preRun,
+    postRun: postRun,
     timeout: timeout,
     verboseTrace: verboseTrace,
     chainStackTraces: chainStackTraces,
@@ -559,6 +567,8 @@ class Configuration {
     skipReason: null,
     testOn: null,
     addTags: null,
+    preRun: null,
+    postRun: null,
   );
 
   /// Suite level configuration that is not allowed in the global test
@@ -575,6 +585,8 @@ class Configuration {
     required Iterable<String>? addTags,
     required bool? allowDuplicateTestNames,
     required bool? allowTestRandomization,
+    Hook? preRun,
+    Hook? postRun,
   }) => Configuration(
     allowDuplicateTestNames: allowDuplicateTestNames,
     allowTestRandomization: allowTestRandomization,
@@ -583,6 +595,8 @@ class Configuration {
     skipReason: skipReason,
     testOn: testOn,
     addTags: addTags,
+    preRun: preRun,
+    postRun: postRun,
     help: null,
     customHtmlTemplatePath: null,
     version: null,
@@ -1082,6 +1096,8 @@ class Configuration {
     BooleanSelector? excludeTags,
     Map<BooleanSelector, SuiteConfiguration>? tags,
     Map<PlatformSelector, SuiteConfiguration>? onPlatform,
+    Hook? preRun,
+    Hook? postRun,
 
     // Test-level configuration
     Timeout? timeout,
@@ -1134,6 +1150,8 @@ class Configuration {
         runtimes: runtimes,
         tags: tags,
         onPlatform: onPlatform,
+        preRun: preRun,
+        postRun: postRun,
         timeout: timeout,
         verboseTrace: verboseTrace,
         chainStackTraces: chainStackTraces,